(#193) feat : 권한 확인

+++ server/src/main/java/seb39_40/coffeewithme/cafe/service/impl/CafePostServiceImpl.java
@@ -1,6 +1,9 @@
 package seb39_40.coffeewithme.cafe.service.impl;
 
 import lombok.RequiredArgsConstructor;
+import org.springframework.http.HttpStatus;
+import org.springframework.security.core.Authentication;
+import org.springframework.security.core.context.SecurityContextHolder;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 import seb39_40.coffeewithme.cafe.domain.Cafe;
@@ -15,11 +18,14 @@ import seb39_40.coffeewithme.image.service.ImageService;
 import seb39_40.coffeewithme.review.domain.Review;
 import seb39_40.coffeewithme.review.service.ReviewService;
 import seb39_40.coffeewithme.tag.service.TagService;
+import seb39_40.coffeewithme.user.domain.User;
+import seb39_40.coffeewithme.user.service.UserService;
 
 import java.util.List;
 import java.util.Objects;
 
 import static seb39_40.coffeewithme.cafe.dto.CafeRequestDto.*;
+import static seb39_40.coffeewithme.common.service.GetUserInfo.getUserId;
 
 @Service
 @RequiredArgsConstructor
@@ -28,6 +34,7 @@ public class CafePostServiceImpl implements CafePostService {
     private final TagService tagService;
     private final ImageService imageService;
     private final ReviewService reviewService;
+    private final UserService userService;
     private final CafeMapper cafeMapper;
 
 
@@ -36,8 +43,12 @@ public class CafePostServiceImpl implements CafePostService {
         Cafe cafe = cafeMapper.cafeDtoToCafe(postDto);
         cafe.setCafeTags(tagService.createCafeTag(cafe, postDto.getTags()));
 
+        User owner = userService.findByEmail(getUserId());
+        cafe.setUser(owner);
+
         // 메인, 메뉴 순으로 이미지 저장
-        if (Objects.equals(postDto.getMainImg(), postDto.getMenuImg())) throw new BusinessLogicException(ExceptionCode.INVALID_INPUT_VALUE);
+        if (Objects.equals(postDto.getMainImg(), postDto.getMenuImg()))
+            throw new BusinessLogicException(HttpStatus.FORBIDDEN, "이미지 입력이 잘못되었습니다.");
         cafe.addImages(imageService.findById(postDto.getMainImg()));
         cafe.addImages(imageService.findById(postDto.getMenuImg()));
         return cafeService.save(cafe);
@@ -46,6 +57,8 @@ public class CafePostServiceImpl implements CafePostService {
     @Transactional
     public void delete(Long id){
         Cafe cafe = cafeService.find(id);
+        if (!cafe.getUser().getEmail().equals(getUserId()))
+            throw new BusinessLogicException(HttpStatus.CONFLICT, "삭제 권한이 없습니다.");
         for (Image image : cafe.getImages()) {image.setCafe(null);}
         List<Review> reviews = cafe.getReviews();
         cafeService.delete(cafe);
@@ -56,11 +69,12 @@ public class CafePostServiceImpl implements CafePostService {
         }
     }
 
-    @Override
+    @Transactional
     public void repost(Long id, Patch patchDto) {
-        // 권한 확인 추가 예정
-
         Cafe origin = cafeService.find(id);
+        if (!origin.getUser().getEmail().equals(getUserId()))
+            throw new BusinessLogicException(HttpStatus.CONFLICT, "수정 권한이 없습니다.");
+
         Cafe target = cafeMapper.cafeDtoToCafe(patchDto);
         origin.updateInformation(target);   // 기본 정보 수정
